aliasable

Struct AliasableMut

pub struct AliasableMut<'a, T: ?Sized> { /* private fields */ }

Basic aliasable alternative to &mut.

Note that this does not circumvent the core aliasing rules of Rust; if you use this to create multiple mutable references to a memory location at the same time, that is still UB. This type just adds a few abilities:

You may hold any number of AliasableMuts and no references to a location.

You may hold any number of AliasableMuts and any number of shared references to a location at once.

You may hold any number of AliasableMuts and one mutable reference to a location at once.

Implementations

impl<'a, T: ?Sized> AliasableMut<'a, T>

pub fn from_unique(ptr: &'a mut T) -> Self

Construct an AliasableMut from an &mut.

pub fn into_unique(ptr: Self) -> &'a mut T

Consumes self and converts it into a non-aliasable &mut.

pub fn into_unique_pin(pin: Pin<Self>) -> Pin<&'a mut T>

Convert a pinned AliasableMut to a pinned &mut.

pub fn from_unique_pin(pin: Pin<&'a mut T>) -> Pin<Self>

Convert a pinned &mut to a pinned AliasableMut.